requirements.txt
pandas>=1.3.0
numpy>=1.20.0
psycopg2-binary>=2.9.0
sqlalchemy>=1.4.0